+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # next
 
 - Add newrelic user to system, so agent will start.
+- Fix attributes for backup based on testing, ES plugin changes
 
 # 3.2.1